Refactor GMP integration and replace mpz_class with custom Integer class
- Removed checks for GMP's C++ interface and replaced mpz_class usage with a custom Integer class defined in gmp_integer.h.
- Updated matrix entry functions to use mpz_srcptr and mpz_ptr instead of mpz_class.
- Modified various source files to accommodate the new Integer class, including changes in conversion functions and API interfaces.
- Enhanced error handling in file reading functions to ensure robustness.
- Updated Makefile configurations to reflect changes in GMP usage and removed unnecessary dependencies on gmpxx.h.

+++ src/groebner/Bounded.cpp
@@ -850,20 +850,20 @@ _4ti2_::lp_weight_l2(
 
     if (basis.get_number() == 0) { return; }
     int index = 0;
-    RationalType dot = Vector::dot(rhs, basis[0]);
+    RationalType dot = DOUBLE(Vector::dot(rhs, basis[0]));
     RationalType ratio = 0;
     for (int j = 0; j < basis.get_size(); ++j)
     {
-        ratio += basis[0][j]*(basis[0][j]/dot);
+        ratio += DOUBLE(basis[0][j])*(DOUBLE(basis[0][j])/dot);
     }
     RationalType min = ratio;
     for (int i = 1; i < basis.get_number(); ++i)
     {
         ratio = 0;
-        dot = Vector::dot(rhs, basis[i]);
+        dot = DOUBLE(Vector::dot(rhs, basis[i]));
         for (int j = 0; j < basis.get_size(); ++j)
         {
-            ratio += basis[i][j]*(basis[i][j]/dot);
+            ratio += DOUBLE(basis[i][j])*(DOUBLE(basis[i][j])/dot);
         }
         if (ratio > min) { index = i; min = ratio; }
     }
